Solution to Check if a Number is Square of Another

+++ Check if a Number is Square of Another@@6788dd63490637cfc44e1ba4/code.c
@@ -5,7 +5,7 @@ int main(){
     int a,b,power;
     printf("");
     scanf("%d%d",&a,&b);
-    power=pow(b,2);
+    power=(int)pow(b,2);
     if(a==power){
         printf("Yes");
     }
